+++ src/extension_core/surfer_document.ts
@@ -75,12 +75,12 @@ export class SurferDocument extends VaporviewDocument implements vscode.CustomDo
     setscopetop: (name: string, id: number, tpe: string) => {
       const scope = createScope(name, tpe, "", id, -1, this.uri);
       this.treeData.push(scope);
-      this._netlistIdTable[id] = { netlistItem: scope, displayedItem: undefined, signalId: 0 };
+      this._netlistIdTable[id] = { netlistItem: scope, signalId: 0 };
     },
     setvartop: (name: string, id: number, signalid: number, tpe: string, encoding: string, width: number, msb: number, lsb: number) => {
       const varItem = createVar(name, tpe, encoding, "", id, signalid, width, msb, lsb, false /*isFsdb*/, this.uri);
       this.treeData.push(varItem);
-      this._netlistIdTable[id] = { netlistItem: varItem, displayedItem: undefined, signalId: signalid };
+      this._netlistIdTable[id] = { netlistItem: varItem, signalId: signalid };
     },
     setmetadata: (scopecount: number, varcount: number, timescale: number, timeunit: string) => {
       this.setMetadata(scopecount, varcount, timescale, timeunit);
@@ -149,7 +149,7 @@ export class SurferDocument extends VaporviewDocument implements vscode.CustomDo
         } else {
           varTable[child.name].push(varItem);
         }
-        this.netlistIdTable[child.netlistId] = { netlistItem: varItem, displayedItem: undefined, signalId: child.signalId };
+        this.netlistIdTable[child.netlistId] = { netlistItem: varItem, signalId: child.signalId };
       });
 
       callLimit--;