+++ CHANGELOG.md
@@ -10,6 +10,7 @@
   - Drag and drop from tree view places signals at the end of group
   - Files with one time table entry now show values
   - New groups now appear where right click event happened
+- Removed "Displayed Signals" View
 
 ## 1.4.1 - 9/12/2025 - Latest Release